refactor: Moved the CreateCheckoutResponse and SavedCardDto classes to the root dto package.
